# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Queue & Appointment SaaS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /businesses/{businessId}/professionals:
    get:
      operationId: listProfessionals
      tags: [professionals]
      summary: List professionals for a business
      parameters:
        - name: businessId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of professionals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Professional"
    post:
      operationId: createProfessional
      tags: [professionals]
      summary: Create a professional
      parameters:
        - name: businessId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProfessionalInput"
      responses:
        "201":
          description: Created professional
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Professional"
  /businesses/{businessId}/professionals/{id}:
    put:
      operationId: updateProfessional
      tags: [professionals]
      summary: Update professional
      parameters:
        - name: businessId
          in: path
          required: true
          schema:
            type: string
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProfessionalInput"
      responses:
        "200":
          description: Updated professional
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Professional"
    delete:
      operationId: deleteProfessional
      tags: [professionals]
      summary: Delete professional
      parameters:
        - name: businessId
          in: path
          required: true
          schema:
            type: string
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted

  # ─── Queues ──────────────────────────────────────────────────────────────────
  /queues:
    get:
      operationId: listQueues
      tags: [queues]
      summary: List queues for a business/date
      parameters:
        - name: businessId
          in: query
          required: true
          schema:
            type: string
        - name: date
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of queues
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Queue"
    post:
      operationId: createQueue
      tags: [queues]
      summary: Open a queue for today
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateQueueInput"
      responses:
        "201":
          description: Created queue
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Queue"
  /queues/{id}:
    get:
      operationId: getQueue
      tags: [queues]
      summary: Get queue details with live entries
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Queue with entries
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QueueWithEntries"
  /queues/{id}/status:
    patch:
      operationId: updateQueueStatus
      tags: [queues]
      summary: Open, pause or close queue
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateQueueStatusInput"
      responses:
        "200":
          description: Updated queue
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Queue"
  /queues/{id}/call-next:
    post:
      operationId: callNext
      tags: [queues]
      summary: Call the next client in queue
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Called entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QueueEntry"

  # ─── Queue Entries ────────────────────────────────────────────────────────────
  /queues/{queueId}/entries:
    get:
      operationId: listQueueEntries
      tags: [queue-entries]
      summary: List entries for a queue
      parameters:
        - name: queueId
          in: path
          required: true
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/QueueEntry"
    post:
      operationId: joinQueue
      tags: [queue-entries]
      summary: Join queue (anonymous)
      parameters:
        - name: queueId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JoinQueueInput"
      responses:
        "201":
          description: Queue entry created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QueueEntry"
  /queues/{queueId}/entries/{id}:
    get:
      operationId: getQueueEntry
      tags: [queue-entries]
      summary: Get queue entry status
      parameters:
        - name: queueId
          in: path
          required: true
          schema:
            type: string
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Queue entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QueueEntry"
    patch:
      operationId: updateQueueEntryStatus
      tags: [queue-entries]
      summary: Update entry status (done, no_show, cancelled)
      parameters:
        - name: queueId
          in: path
          required: true
          schema:
            type: string
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateEntryStatusInput"
      responses:
        "200":
          description: Updated entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QueueEntry"

  # ─── Appointments ─────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Professional:
      type: object
      properties:
        id:
          type: string
        tenantId:
          type: string
        businessId:
          type: string
        clerkId:
          type: string
          nullable: true
        name:
          type: string
        role:
          type: string
        avatarUrl:
          type: string
          nullable: true
        isActive:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, tenantId, businessId, name, role, isActive, createdAt, updatedAt]
    CreateProfessionalInput:
      type: object
      properties:
        name:
          type: string
        role:
          type: string
        clerkId:
          type: string
          nullable: true
        avatarUrl:
          type: string
          nullable: true
        isActive:
          type: boolean
      required: [name]

    # ─── Queues ───
    Queue:
      type: object
      properties:
        id:
          type: string
        tenantId:
          type: string
        businessId:
          type: string
        serviceId:
          type: string
          nullable: true
        professionalId:
          type: string
          nullable: true
        date:
          type: string
        status:
          type: string
          enum: [open, paused, closed]
        currentTicket:
          type: integer
        lastTicket:
          type: integer
        avgWaitMinutes:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, tenantId, businessId, date, status, currentTicket, lastTicket, avgWaitMinutes, createdAt, updatedAt]
    QueueWithEntries:
      allOf:
        - $ref: "#/components/schemas/Queue"
        - type: object
          properties:
            entries:
              type: array
              items:
                $ref: "#/components/schemas/QueueEntry"
          required: [entries]
    CreateQueueInput:
      type: object
      properties:
        businessId:
          type: string
        serviceId:
          type: string
          nullable: true
        professionalId:
          type: string
          nullable: true
        date:
          type: string
      required: [businessId, date]
    UpdateQueueStatusInput:
      type: object
      properties:
        status:
          type: string
          enum: [open, paused, closed]
      required: [status]

    # ─── Queue Entries ───
    QueueEntry:
      type: object
      properties:
        id:
          type: string
        tenantId:
          type: string
        queueId:
          type: string
        ticketNumber:
          type: integer
        clientName:
          type: string
        clientPhone:
          type: string
          nullable: true
        serviceId:
          type: string
          nullable: true
        professionalId:
          type: string
          nullable: true
        status:
          type: string
          enum: [waiting, called, in_service, done, cancelled, no_show]
        calledAt:
          type: string
          nullable: true
        servedAt:
          type: string
          nullable: true
        finishedAt:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
        updatedAt:
          type: string
        waitingAhead:
          type: integer
          nullable: true
        estimatedWaitMinutes:
          type: integer
          nullable: true
      required: [id, tenantId, queueId, ticketNumber, clientName, status, createdAt, updatedAt]
    JoinQueueInput:
      type: object
      properties:
        clientName:
          type: string
        clientPhone:
          type: string
          nullable: true
        serviceId:
          type: string
          nullable: true
        professionalId:
          type: string
          nullable: true
      required: [clientName]
    UpdateEntryStatusInput:
      type: object
      properties:
        status:
          type: string
          enum: [waiting, called, in_service, done, cancelled, no_show]
        notes:
          type: string
          nullable: true
      required: [status]

    # ─── Appointments ───
...